merge
add

monica's bugfix

gigi

+++ CalendarFXView/src/main/java/com/calendarfx/view/WeekView.java
@@ -250,7 +250,7 @@ public class WeekView extends DayViewBase {
     private void updateStartAndEndDates() {
         LocalDate date = calculateStartDate();
         startDate.set(date);
-        endDate.set(date.plusDays(getNumberOfDays() - 1));
+        endDate.set(date.plusDays(getNumberOfDays() - (long)1));
     }
 
     private LocalDate calculateStartDate() {
@@ -399,6 +399,7 @@ public class WeekView extends DayViewBase {
 
             @Override
             public void setValue(Object value) {
+                //empty mehtod
             }
 
             @Override
@@ -436,6 +437,7 @@ public class WeekView extends DayViewBase {
 
             @Override
             public void setValue(Object value) {
+                //empty method
             }
 
             @Override
